Update index.html
Added some improvements to the GUI. Now the rectangular shape allows you to specify number of rows and/or columns. Useful for creating areas for strimers and matrix panels. Added option to lock the object position and size. Added button for generating the full html file by taking the code from the GitHub repository and replacing the header. It works, but I need to update the class to handle some of the new stuff in the EffectBuilder.