Update index.html
Current values are now stored to the default value in the output script.

Zip file generation accounts for the user's time and time zone so that the files inside the zip has a timestamp that matches the zip file.

Partial fix to resizing of ring objects. It is still buggy but it works.

fixed draging of multiple objects so that it stops as soon as the biggest object touches a border. Also, dragging is now stopped when the mouse gets out of the canvas.

Added thirdparty open source libraries to the About modal.

Fixed configStore so that values are updated correctly.