Resolution changes
Now displaying all values in the SignalRGB default 320x200 resolution. Exported file will be done at that resolution now. This will be more intuitive going forward as the users are used to work in 320x200. However, the display in the website will be 4X larger (1280x800) to improve image quality.